backend/requirements.txt
# FastAPI
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart

# Database
sqlalchemy[asyncio]
asyncpg
redis

# Vector Store
qdrant-client

# Embeddings
httpx

# LangChain
langchain
langchain-core
langchain-text-splitters
# langchain-google-genai
# langchain-openai
# langchain-anthropic
langchain-ollama

# Document Processing
python-docx
pymupdf
aiofiles

# Memory
mem0ai
